+++ packages/solidart/lib/src/core/read_signal.dart
@@ -33,10 +33,6 @@ class ReadSignal<T> extends Atom implements SignalBase<T> {
 
   final T _value;
 
-  /// Whether or not there has already been a previous value. It is especially
-  /// helpful if [T] is nullable.
-  bool hasPreviousValue = false;
-
   /// All the observers
   @internal
   final List<ObserveCallback<T>> listeners = [];
@@ -49,6 +45,9 @@ class ReadSignal<T> extends Atom implements SignalBase<T> {
 
   @override
   T call() => value;
+
+  @override
+  bool get hasPreviousValue => false;
 
   // coverage:ignore-start
   @override